requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.3
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
wincertstore==0.2
